Move `NetworkRequestIndicator` directly into `components` directory
