+++ src/components/NetworkRequestIndicator/NetworkRequestIndicator.module.scss
