+++ src/components/NetworkRequestIndicator/NetworkRequestIndicator.tsx
